services/api/requirements.txt
boto3>=1.34.0
openai>=1.12.0
PyJWT>=2.8.0
requests>=2.31.0
